+++ CHANGELOG.md
@@ -3,8 +3,9 @@
 ## next
 
 - Add a new option: `initialCoverage` (#526).
+- Support for providing pivot pointer coordinates to `zoom` and `zoomTo` methods (#202).
+- Don't override the padding right of the body when the width of the scrollbar is zero (#197).
 - Add keyboard support to playing view (#90).
-- Don't override the padding right of the body when the width of the scrollbar is zero.
 
 ## 1.10.5 (Apr 5, 2022)