+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## next
+## 1.11.0 (Oct 16, 2022)
 
 - Add a new option: `initialCoverage` (#314, #526).
 - Don't load images in the list when hide the navbar (#451).